backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
qrcode[pil]==7.4.2
fpdf2==2.7.9
